@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.31.0
-// 	protoc        v4.23.3
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.27.3
 // source: grpc-proto/message.proto
 
 package grpc_proto
@@ -1983,7 +1983,7 @@ func file_grpc_proto_message_proto_rawDescGZIP() []byte {
 }
 
 var file_grpc_proto_message_proto_msgTypes = make([]protoimpl.MessageInfo, 29)
-var file_grpc_proto_message_proto_goTypes = []interface{}{
+var file_grpc_proto_message_proto_goTypes = []any{
 	(*PingApiMessageRequest)(nil),             // 0: message.PingApiMessageRequest
 	(*PingApiMessageResponse)(nil),            // 1: message.PingApiMessageResponse
 	(*PingTasksQueueRequest)(nil),             // 2: message.PingTasksQueueRequest
@@ -2063,7 +2063,7 @@ func file_grpc_proto_message_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_grpc_proto_message_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2075,7 +2075,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*PingApiMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2087,7 +2087,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[2].Exporter = func(v any, i int) any {
 			switch v := v.(*PingTasksQueueRequest); i {
 			case 0:
 				return &v.state
@@ -2099,7 +2099,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[3].Exporter = func(v any, i int) any {
 			switch v := v.(*PingTasksQueueResponse); i {
 			case 0:
 				return &v.state
@@ -2111,7 +2111,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[4].Exporter = func(v any, i int) any {
 			switch v := v.(*CustomerMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2123,7 +2123,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[5].Exporter = func(v any, i int) any {
 			switch v := v.(*CustomerMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2135,7 +2135,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[6].Exporter = func(v any, i int) any {
 			switch v := v.(*UserSendMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2147,7 +2147,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[7].Exporter = func(v any, i int) any {
 			switch v := v.(*UserSendMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2159,7 +2159,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[8].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[8].Exporter = func(v any, i int) any {
 			switch v := v.(*EditMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2171,7 +2171,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[9].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[9].Exporter = func(v any, i int) any {
 			switch v := v.(*EditMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2183,7 +2183,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[10].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[10].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2195,7 +2195,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[11].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[11].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2207,7 +2207,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[12].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[12].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2219,7 +2219,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[13].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[13].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2231,7 +2231,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[14].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[14].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchConversationMessagesRequest); i {
 			case 0:
 				return &v.state
@@ -2243,7 +2243,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[15].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[15].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchConversationMessagesResponse); i {
 			case 0:
 				return &v.state
@@ -2255,7 +2255,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[16].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[16].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteConversationRequest); i {
 			case 0:
 				return &v.state
@@ -2267,7 +2267,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[17].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[17].Exporter = func(v any, i int) any {
 			switch v := v.(*DeleteConversationResponse); i {
 			case 0:
 				return &v.state
@@ -2279,7 +2279,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[18].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[18].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchAllUserConversationsRequest); i {
 			case 0:
 				return &v.state
@@ -2291,7 +2291,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[19].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[19].Exporter = func(v any, i int) any {
 			switch v := v.(*FetchAllUserConversationsResponse); i {
 			case 0:
 				return &v.state
@@ -2303,7 +2303,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[20].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[20].Exporter = func(v any, i int) any {
 			switch v := v.(*ChatbotMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2315,7 +2315,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[21].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[21].Exporter = func(v any, i int) any {
 			switch v := v.(*ChatbotMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2327,7 +2327,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[22].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[22].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateTaskMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2339,7 +2339,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[23].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[23].Exporter = func(v any, i int) any {
 			switch v := v.(*CreateTaskMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2351,7 +2351,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[24].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[24].Exporter = func(v any, i int) any {
 			switch v := v.(*CompleteTaskMessageRequest); i {
 			case 0:
 				return &v.state
@@ -2363,7 +2363,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[25].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[25].Exporter = func(v any, i int) any {
 			switch v := v.(*CompleteTaskMessageResponse); i {
 			case 0:
 				return &v.state
@@ -2375,7 +2375,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[26].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[26].Exporter = func(v any, i int) any {
 			switch v := v.(*Conversation); i {
 			case 0:
 				return &v.state
@@ -2387,7 +2387,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[27].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[27].Exporter = func(v any, i int) any {
 			switch v := v.(*Message); i {
 			case 0:
 				return &v.state
@@ -2399,7 +2399,7 @@ func file_grpc_proto_message_proto_init() {
 				return nil
 			}
 		}
-		file_grpc_proto_message_proto_msgTypes[28].Exporter = func(v interface{}, i int) interface{} {
+		file_grpc_proto_message_proto_msgTypes[28].Exporter = func(v any, i int) any {
 			switch v := v.(*TaskData); i {
 			case 0:
 				return &v.state
