@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.23.3
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.27.3
 // source: grpc-proto/behaviour.proto
 
 package grpc_proto
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	BehaviourService_PingApiBehaviour_FullMethodName = "/behaviour.BehaviourService/PingApiBehaviour"
@@ -40,8 +40,9 @@ func NewBehaviourServiceClient(cc grpc.ClientConnInterface) BehaviourServiceClie
 }
 
 func (c *behaviourServiceClient) PingApiBehaviour(ctx context.Context, in *PingApiBehaviourRequest, opts ...grpc.CallOption) (*PingApiBehaviourResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingApiBehaviourResponse)
-	err := c.cc.Invoke(ctx, BehaviourService_PingApiBehaviour_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BehaviourService_PingApiBehaviour_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -49,8 +50,9 @@ func (c *behaviourServiceClient) PingApiBehaviour(ctx context.Context, in *PingA
 }
 
 func (c *behaviourServiceClient) BehaviourFetch(ctx context.Context, in *BehaviourFetchRequest, opts ...grpc.CallOption) (*BehaviourFetchResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(BehaviourFetchResponse)
-	err := c.cc.Invoke(ctx, BehaviourService_BehaviourFetch_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, BehaviourService_BehaviourFetch_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -59,16 +61,19 @@ func (c *behaviourServiceClient) BehaviourFetch(ctx context.Context, in *Behavio
 
 // BehaviourServiceServer is the server API for BehaviourService service.
 // All implementations must embed UnimplementedBehaviourServiceServer
-// for forward compatibility
+// for forward compatibility.
 type BehaviourServiceServer interface {
 	PingApiBehaviour(context.Context, *PingApiBehaviourRequest) (*PingApiBehaviourResponse, error)
 	BehaviourFetch(context.Context, *BehaviourFetchRequest) (*BehaviourFetchResponse, error)
 	mustEmbedUnimplementedBehaviourServiceServer()
 }
 
-// UnimplementedBehaviourServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedBehaviourServiceServer struct {
-}
+// UnimplementedBehaviourServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedBehaviourServiceServer struct{}
 
 func (UnimplementedBehaviourServiceServer) PingApiBehaviour(context.Context, *PingApiBehaviourRequest) (*PingApiBehaviourResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method PingApiBehaviour not implemented")
@@ -77,6 +82,7 @@ func (UnimplementedBehaviourServiceServer) BehaviourFetch(context.Context, *Beha
 	return nil, status.Errorf(codes.Unimplemented, "method BehaviourFetch not implemented")
 }
 func (UnimplementedBehaviourServiceServer) mustEmbedUnimplementedBehaviourServiceServer() {}
+func (UnimplementedBehaviourServiceServer) testEmbeddedByValue()                          {}
 
 // UnsafeBehaviourServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to BehaviourServiceServer will
@@ -86,6 +92,13 @@ type UnsafeBehaviourServiceServer interface {
 }
 
 func RegisterBehaviourServiceServer(s grpc.ServiceRegistrar, srv BehaviourServiceServer) {
+	// If the following call pancis, it indicates UnimplementedBehaviourServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&BehaviourService_ServiceDesc, srv)
 }
 
