@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.23.3
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.27.3
 // source: grpc-proto/message.proto
 
 package grpc_proto
@@ -15,8 +15,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	MessageService_PingApiMessage_FullMethodName            = "/message.MessageService/PingApiMessage"
@@ -62,8 +62,9 @@ func NewMessageServiceClient(cc grpc.ClientConnInterface) MessageServiceClient {
 }
 
 func (c *messageServiceClient) PingApiMessage(ctx context.Context, in *PingApiMessageRequest, opts ...grpc.CallOption) (*PingApiMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingApiMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_PingApiMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_PingApiMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -71,8 +72,9 @@ func (c *messageServiceClient) PingApiMessage(ctx context.Context, in *PingApiMe
 }
 
 func (c *messageServiceClient) CustomerMessage(ctx context.Context, in *CustomerMessageRequest, opts ...grpc.CallOption) (*CustomerMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CustomerMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_CustomerMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_CustomerMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -80,8 +82,9 @@ func (c *messageServiceClient) CustomerMessage(ctx context.Context, in *Customer
 }
 
 func (c *messageServiceClient) ChatbotMessage(ctx context.Context, in *ChatbotMessageRequest, opts ...grpc.CallOption) (*ChatbotMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ChatbotMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_ChatbotMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_ChatbotMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -89,8 +92,9 @@ func (c *messageServiceClient) ChatbotMessage(ctx context.Context, in *ChatbotMe
 }
 
 func (c *messageServiceClient) PingTasksQueue(ctx context.Context, in *PingTasksQueueRequest, opts ...grpc.CallOption) (*PingTasksQueueResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PingTasksQueueResponse)
-	err := c.cc.Invoke(ctx, MessageService_PingTasksQueue_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_PingTasksQueue_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -98,8 +102,9 @@ func (c *messageServiceClient) PingTasksQueue(ctx context.Context, in *PingTasks
 }
 
 func (c *messageServiceClient) CreateTaskMessage(ctx context.Context, in *CreateTaskMessageRequest, opts ...grpc.CallOption) (*CreateTaskMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateTaskMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_CreateTaskMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_CreateTaskMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -107,8 +112,9 @@ func (c *messageServiceClient) CreateTaskMessage(ctx context.Context, in *Create
 }
 
 func (c *messageServiceClient) CompleteTaskMessage(ctx context.Context, in *CompleteTaskMessageRequest, opts ...grpc.CallOption) (*CompleteTaskMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CompleteTaskMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_CompleteTaskMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_CompleteTaskMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -116,8 +122,9 @@ func (c *messageServiceClient) CompleteTaskMessage(ctx context.Context, in *Comp
 }
 
 func (c *messageServiceClient) UserSendMessage(ctx context.Context, in *UserSendMessageRequest, opts ...grpc.CallOption) (*UserSendMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(UserSendMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_UserSendMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_UserSendMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -125,8 +132,9 @@ func (c *messageServiceClient) UserSendMessage(ctx context.Context, in *UserSend
 }
 
 func (c *messageServiceClient) EditMessage(ctx context.Context, in *EditMessageRequest, opts ...grpc.CallOption) (*EditMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(EditMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_EditMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_EditMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -134,8 +142,9 @@ func (c *messageServiceClient) EditMessage(ctx context.Context, in *EditMessageR
 }
 
 func (c *messageServiceClient) FetchMessage(ctx context.Context, in *FetchMessageRequest, opts ...grpc.CallOption) (*FetchMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FetchMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_FetchMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_FetchMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -143,8 +152,9 @@ func (c *messageServiceClient) FetchMessage(ctx context.Context, in *FetchMessag
 }
 
 func (c *messageServiceClient) DeleteMessage(ctx context.Context, in *DeleteMessageRequest, opts ...grpc.CallOption) (*DeleteMessageResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteMessageResponse)
-	err := c.cc.Invoke(ctx, MessageService_DeleteMessage_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_DeleteMessage_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -152,8 +162,9 @@ func (c *messageServiceClient) DeleteMessage(ctx context.Context, in *DeleteMess
 }
 
 func (c *messageServiceClient) FetchConversationMessages(ctx context.Context, in *FetchConversationMessagesRequest, opts ...grpc.CallOption) (*FetchConversationMessagesResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FetchConversationMessagesResponse)
-	err := c.cc.Invoke(ctx, MessageService_FetchConversationMessages_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_FetchConversationMessages_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -161,8 +172,9 @@ func (c *messageServiceClient) FetchConversationMessages(ctx context.Context, in
 }
 
 func (c *messageServiceClient) DeleteConversation(ctx context.Context, in *DeleteConversationRequest, opts ...grpc.CallOption) (*DeleteConversationResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteConversationResponse)
-	err := c.cc.Invoke(ctx, MessageService_DeleteConversation_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_DeleteConversation_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -170,8 +182,9 @@ func (c *messageServiceClient) DeleteConversation(ctx context.Context, in *Delet
 }
 
 func (c *messageServiceClient) FetchAllUserConversations(ctx context.Context, in *FetchAllUserConversationsRequest, opts ...grpc.CallOption) (*FetchAllUserConversationsResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(FetchAllUserConversationsResponse)
-	err := c.cc.Invoke(ctx, MessageService_FetchAllUserConversations_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, MessageService_FetchAllUserConversations_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -180,7 +193,7 @@ func (c *messageServiceClient) FetchAllUserConversations(ctx context.Context, in
 
 // MessageServiceServer is the server API for MessageService service.
 // All implementations must embed UnimplementedMessageServiceServer
-// for forward compatibility
+// for forward compatibility.
 type MessageServiceServer interface {
 	PingApiMessage(context.Context, *PingApiMessageRequest) (*PingApiMessageResponse, error)
 	CustomerMessage(context.Context, *CustomerMessageRequest) (*CustomerMessageResponse, error)
@@ -198,9 +211,12 @@ type MessageServiceServer interface {
 	mustEmbedUnimplementedMessageServiceServer()
 }
 
-// UnimplementedMessageServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedMessageServiceServer struct {
-}
+// UnimplementedMessageServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedMessageServiceServer struct{}
 
 func (UnimplementedMessageServiceServer) PingApiMessage(context.Context, *PingApiMessageRequest) (*PingApiMessageResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method PingApiMessage not implemented")
@@ -242,6 +258,7 @@ func (UnimplementedMessageServiceServer) FetchAllUserConversations(context.Conte
 	return nil, status.Errorf(codes.Unimplemented, "method FetchAllUserConversations not implemented")
 }
 func (UnimplementedMessageServiceServer) mustEmbedUnimplementedMessageServiceServer() {}
+func (UnimplementedMessageServiceServer) testEmbeddedByValue()                        {}
 
 // UnsafeMessageServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to MessageServiceServer will
@@ -251,6 +268,13 @@ type UnsafeMessageServiceServer interface {
 }
 
 func RegisterMessageServiceServer(s grpc.ServiceRegistrar, srv MessageServiceServer) {
+	// If the following call pancis, it indicates UnimplementedMessageServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&MessageService_ServiceDesc, srv)
 }
 
